+++ packages/qw-element/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.2.5] - 19/07/2021
+
+### Added
+
+- getAllNextSiblings function
+- getAllPreviousSiblings function
+
 ## [0.2.4] - 19/07/2021
 
 ### Added